Notification: Move bongo related methods to notification_bongo
